backend/requirements.txt
amqp==2.3.2
Babel==2.6.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-appconf==1.0.2
django-cors-headers==2.4.0
django-extensions==2.1.4
django-imagekit==4.0.2
django-js-asset==1.1.0
django-mptt==0.9.1
django-rest-elasticsearch==0.4.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
flower==0.9.2
idna==2.8
kombu==4.2.2.post1
pilkit==2.0
Pillow==5.4.1
psycopg2==2.7.6.1
PyJWT==1.7.1
pytest-runner==4.2
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
six==1.12.0
tornado==5.1.1
transliterate==1.10.2
urllib3==1.24.1
uWSGI==2.0.17.1
vine==1.2.0
